feat(matomo-analytics): export apifactory for dynamic plugin
