+++ plugins/analytics-module-matomo/app-config.janus-idp.yaml
@@ -1,0 +1,5 @@
+app:
+  analytics:
+    matomo:
+      host: ${ANALYTICS_MATOMO_INSTANCE_URL}
+      siteId: ${ANALYTICS_MATOMO_SITE_ID}
