+++ plugins/analytics-module-matomo/package.json
@@ -10,11 +10,12 @@
     "types": "dist/index.d.ts"
   },
   "backstage": {
-    "role": "frontend-plugin-module"
+    "role": "frontend-plugin"
   },
   "scripts": {
     "start": "backstage-cli package start",
     "build": "backstage-cli package build",
+    "export-dynamic": "janus-cli package export-dynamic-plugin",
     "tsc": "tsc",
     "lint": "backstage-cli package lint",
     "test": "backstage-cli package test --passWithNoTests --coverage",
@@ -33,6 +34,7 @@
     "@backstage/core-app-api": "1.12.0",
     "@backstage/dev-utils": "1.0.27",
     "@backstage/test-utils": "1.5.0",
+    "@janus-idp/cli": "1.7.5",
     "@testing-library/jest-dom": "6.4.2",
     "@testing-library/react": "14.2.1",
     "@testing-library/user-event": "14.5.1",
@@ -42,7 +44,9 @@
   },
   "files": [
     "dist",
-    "config.d.ts"
+    "dist-scalprum",
+    "config.d.ts",
+    "app-config.janus-idp.yaml"
   ],
   "configSchema": "config.d.ts"
 }
